+++ contrib/platform/src/com/sun/jna/platform/win32/COM/util/IConnectionPoint.java
@@ -24,7 +24,9 @@
 package com.sun.jna.platform.win32.COM.util;
 
 import com.sun.jna.platform.win32.COM.COMException;
+import com.sun.jna.platform.win32.COM.util.annotation.ComInterface;
 
+@ComInterface(iid="{B196B284-BAB4-101A-B69C-00AA00341D07}")
 public interface IConnectionPoint {
 
 	/**